Extract ReviewRequestButton (with modal)
